Memory dump pane update: updating memory is now very very efficient.
Bad news: Highlighting bytes is much less efficient.
Good news: Highlighting bytes happens very little, and no cpu time is
wasted on highlighting while the dump pane is hidden.

All in all, I believe it's working about as well as it's going to (sans
a brilliant idea for how to highlight bytes without having a lot of code
that does almost-the-same-exact-thing-as-the-other - in any case, it's a
lot of cursor movement, which is expensive).

There is a slight scrollbar issue that I noticed, and I'm not entirely
how to fix it, but I imagine we'll deal with the same issue in our
upcoming implementation of the OS/loader trace.


git-svn-id: http://pep8-1.googlecode.com/svn/trunk@237 0efa5a84-20a4-11de-9909-f37ff7dbbc12