You can now double-click .pep and .pepo files in the filesystem to open them! ...at least, on Mac. Haven't begun to try to do this on PC, but it might work the same way after I associate those filetypes with the app.
Either way, test it out with the figures folder! Trying to associate a
unique icon with the pep documents as well, but no luck so far.

Fixed a save prompt bug that occured on quitting. Added a .plist file so
we aren't relying on a generated one for our application info. We'll
probably want to modify some of the fields in there at some point.


git-svn-id: http://pep8-1.googlecode.com/svn/trunk@242 0efa5a84-20a4-11de-9909-f37ff7dbbc12