The memory trace pane will now update on the fly for globals, except for globals that are modified by trap instructions.
It will also sometimes crash right now when running more than one
program per-run of the application (for example, fig5.26). However, it
does handle all types of trace tags.

It does not yet nicely handle global arrays, however.

Also implemented a double-click-on-the-label-to-expand feature on the
trace tab.

Anyways, check it out! Fig6.12 is a good example, as is 5.21.

(Are we clearing the tables we're using for the trace tags between runs?
That could be causing our crashes...)


git-svn-id: http://pep8-1.googlecode.com/svn/trunk@341 0efa5a84-20a4-11de-9909-f37ff7dbbc12