Highlight of all allowed characters in identifiers
This change is purported to make all allowed characters in identifiers to receive the proper highlighting. Previously, an identifier with an accented character would only be highlighted after the accented character, creating a distraction when reading the code.

The replacement was copied from line 32 if asm.cpp. It is my understanding that it is that line that decides which characters are allowed in identifiers.
asm.cpp 32: QRegExp Asm::rxIdentifier("^((([A-Z|a-z|_]{1})(\\w*))(:){0,1})");

The first solution was to write "." before the "+", effectively highlighting _all_ characters before the colon. This solution was rejected in favor of the current one in order to provide immediate feedback visual when an invalid character is written in an identifier.

This change was lightly tested at home and apparently worked. "Apparently" means that it seems to work fine, but I don't know if something else broke somewhere and I am not confident in my capacities to properly test this.

The tests were made by compiling in Qt Creator 4.8.1 with the Qt 5.12.1 MSVC2017 64bit kit.